Implemented the ability to add new parts to the system database completely. (Need to adjust the GUI so that the Part ID field can be inputted. Also need to add alerts for when values that must be integers but have not been inputted as such.)
Implemented the ability to remove a selected part from the system database partially.(Need to implement the ability to search for particular parts and adjust the List View to be more presentable.)

Implemented the ability to adjust the stock of a selected part from the system database partially. (Need to implement the ability to search for particular parts and send the data for the selected part to the next scene.)

Made minor corrections to domain classes, DBConnectively class and .fxml files.